fix: Cypress config
Move the RenderSearchParams logic/ command (only used for CT but being compiled during e2e tests ) out of the commands file to prevent the e2e run compilation error.
